moved fetchng funtion in context
